# api/openapi.yaml
openapi: 3.1.0
info:
  title: 神州 HR 考勤系统 API
  version: 1.9.0-wave3
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    V1.9 WAVE-3 契约：在完整保留 WAVE-1 本地账号、登录会话、权限、审计与
    版本化规则配置底座，以及 WAVE-2 组织、员工期初导入与本地人员权威的基础上，
    增加考勤地点、考勤组、员工期限分配、不可变班次版本、工作日历和基础策略绑定。
    所有 ID 均按 string 传输；响应和示例不包含密码、密码哈希、session token 或 reset token。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: 使用本地账号密码登录
      security: []
      x-rate-limit: LOGIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功；session 只通过安全 Cookie 建立
          headers:
            Set-Cookie:
              schema: { type: string }
              description: Host-only、HttpOnly、Path=/、SameSite=Lax；HTTPS 环境必须 Secure。
            X-CSRF-TOKEN:
              schema: { type: string }
              description: 后续状态变更请求使用的 CSRF token，不得持久化到浏览器存储。
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
        '401':
          $ref: '#/components/responses/InvalidCredentials'
        '423':
          $ref: '#/components/responses/AccountLocked'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/session:
    get:
      tags: [Authentication]
      operationId: getCurrentSession
      summary: 获取当前安全会话、能力和服务端菜单
      responses:
        '200':
          description: 当前会话
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
            X-CSRF-TOKEN:
              schema: { type: string }
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: 撤销当前会话并退出
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: 当前会话已撤销
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
  /auth/password/change:
    post:
      tags: [Authentication]
      operationId: changePassword
      summary: 修改当前账号密码并撤销其他会话
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '204':
          description: 密码已修改；除当前重新建立的会话外，旧会话全部失效
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
  /auth/password/first-change:
    post:
      tags: [Authentication]
      operationId: completeFirstPasswordChange
      summary: 完成首次登录强制改密
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '204':
          description: 首次改密完成并撤销旧会话
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/password-reset-requests:
    post:
      tags: [Authentication]
      operationId: requestPasswordReset
      summary: 发起不暴露账号存在性的密码恢复请求
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '202':
          description: 请求统一受理；无论账号是否存在均返回相同响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOperation'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/password-resets:
    post:
      tags: [Authentication]
      operationId: resetPassword
      summary: 使用一次性授权重置密码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetCompletionRequest'
      responses:
        '204':
          description: 密码已重置；授权与旧会话全部失效
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/InvalidResetGrant'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/sessions/{sessionId}/revoke:
    post:
      tags: [Authentication]
      operationId: revokeSession
      summary: 撤销本人会话或具有权限的账号会话
      x-capability: ACCOUNT:EDIT
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '204':
          description: 会话已撤销
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
...
  /access/accounts:
    get:
      tags: [Access]
      operationId: listAccounts
      summary: 分页查询账号
      x-capability: ACCOUNT:READ
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - name: query
          in: query
          schema: { type: string, maxLength: 100 }
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AccountStatus'
      responses:
        '200':
          description: 授权范围内账号
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPage'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
    post:
      tags: [Access]
      operationId: createAccount
      summary: 创建本地账号
      x-capability: ACCOUNT:CREATE
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreateRequest'
      responses:
        '201':
          description: 账号已创建；响应不返回初始密码
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /access/roles:
    get:
      tags: [Access]
      operationId: listRoles
      summary: 查询角色、能力和数据范围定义
      x-capability: ROLE:READ
      responses:
        '200':
          description: 角色列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleView'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SHENZHOUHR_SESSION
      description: 本系统本地账号安全会话；开发身份头不属于正式认证路径。
  headers:
    CorrelationId:
      schema:
        type: string
        minLength: 1
        maxLength: 64
...
  parameters:
    CsrfToken:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema: { type: string, minLength: 32, maxLength: 256 }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum: [createdAt,updatedAt,username,status,occurredAt,templateCode,versionNumber]
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
...
  schemas:
    ResourceId:
      type: string
      minLength: 1
      maxLength: 128
      description: 精确资源标识；禁止转换为 JavaScript Number。
...
    AccountStatus:
      type: string
      enum: [ACTIVE, DISABLED, LOCKED]
    SessionStatus:
      type: string
      enum: [ACTIVE, EXPIRED, REVOKED]
...
    MenuItem:
      type: object
      additionalProperties: false
      required: [key, label, path]
      properties:
        key: { type: string, maxLength: 64 }
        label: { type: string, maxLength: 64 }
        path: { type: string, pattern: '^/' }
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 128 }
        password: { type: string, minLength: 1, maxLength: 256, writeOnly: true }
    SessionView:
      type: object
      additionalProperties: false
      required:
        [sessionId, accountId, username, displayName, status, firstPasswordChangeRequired,
         issuedAt, idleExpiresAt, absoluteExpiresAt, capabilities, menu]
      properties:
        sessionId: { $ref: '#/components/schemas/ResourceId' }
        accountId: { $ref: '#/components/schemas/ResourceId' }
        username: { type: string }
        displayName: { type: string }
        status: { $ref: '#/components/schemas/SessionStatus' }
        firstPasswordChangeRequired: { type: boolean }
        issuedAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        idleExpiresAt: { type: string, format: date-time }
        absoluteExpiresAt: { type: string, format: date-time }
        capabilities:
          type: array
          uniqueItems: true
          items: { type: string }
        menu:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
    PasswordChangeRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1, maxLength: 256, writeOnly: true }
        newPassword: { type: string, minLength: 12, maxLength: 256, writeOnly: true }
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [username]
      properties:
        username: { type: string, minLength: 1, maxLength: 128 }
    PasswordResetCompletionRequest:
      type: object
      additionalProperties: false
      required: [grant, newPassword]
      properties:
        grant: { type: string, minLength: 32, maxLength: 512, writeOnly: true }
        newPassword: { type: string, minLength: 12, maxLength: 256, writeOnly: true }
    AcceptedOperation:
      type: object
      additionalProperties: false
      required: [accepted, correlationId]
      properties:
        accepted: { type: boolean, const: true }
        correlationId: { type: string }
    ReasonRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 2, maxLength: 500 }
    AccountSummary:
      type: object
      additionalProperties: false
      required: [accountId, username, displayName, status, firstPasswordChangeRequired, rowVersion]
      properties:
        accountId: { $ref: '#/components/schemas/ResourceId' }
        username: { type: string }
        displayName: { type: string }
        status: { $ref: '#/components/schemas/AccountStatus' }
        firstPasswordChangeRequired: { type: boolean }
        resetPending: { type: boolean }
        lockedUntil: { type: [string, 'null'], format: date-time }
        lastLoginAt: { type: [string, 'null'], format: date-time }
        rowVersion: { type: integer, format: int64, minimum: 0 }
    AccountDetail:
      allOf:
        - $ref: '#/components/schemas/AccountSummary'
        - type: object
          required: [roles, sessions]
          properties:
            roles:
              type: array
              items:
                $ref: '#/components/schemas/RoleAssignmentView'
            sessions:
              type: array
              items:
                $ref: '#/components/schemas/SessionSummary'
    AccountPage:
      type: object
      additionalProperties: false
      required: [items, total, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccountSummary'
        total: { type: integer, format: int64, minimum: 0 }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
    AccountCreateRequest:
      type: object
      additionalProperties: false
      required: [username, displayName, temporaryPassword, roleAssignments]
      properties:
        username: { type: string, minLength: 3, maxLength: 128 }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        temporaryPassword: { type: string, minLength: 12, maxLength: 256, writeOnly: true }
        roleAssignments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RoleAssignmentRequest'
...
    SessionSummary:
      type: object
      additionalProperties: false
      required: [sessionId, status, createdAt, idleExpiresAt, absoluteExpiresAt]
      properties:
        sessionId: { $ref: '#/components/schemas/ResourceId' }
        status: { $ref: '#/components/schemas/SessionStatus' }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        idleExpiresAt: { type: string, format: date-time }
        absoluteExpiresAt: { type: string, format: date-time }
        revokedAt: { type: [string, 'null'], format: date-time }
        revocationReason: { type: [string, 'null'] }
    RoleView:
      type: object
      additionalProperties: false
      required: [roleId, code, name, capabilities]
      properties:
        roleId: { $ref: '#/components/schemas/ResourceId' }
        code: { type: string }
        name: { type: string }
        capabilities:
          type: array
          uniqueItems: true
          items: { type: string }
    RoleAssignmentRequest:
      type: object
      additionalProperties: false
      required: [roleId, scopeType, validFrom]
      properties:
        roleId: { $ref: '#/components/schemas/ResourceId' }
        scopeType: { type: string, enum: [LEGAL_ENTITY, ORGANIZATION, SELF] }
        scopeResourceId:
          oneOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        validFrom: { type: string, format: date-time }
        validTo: { type: [string, 'null'], format: date-time }
    RoleAssignmentView:
      allOf:
        - $ref: '#/components/schemas/RoleAssignmentRequest'
        - type: object
          required: [assignmentId, roleCode, roleName]
          properties:
            assignmentId: { $ref: '#/components/schemas/ResourceId' }
            roleCode: { type: string }
            roleName: { type: string }
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, retryable]
      properties:
        code: { type: string }
        message: { type: string }
        correlationId: { type: string }
        retryable: { type: boolean }
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        details:
          type: object
          additionalProperties: true
  responses:
    AuthenticationRequired:
      description: 身份无效、会话到期或会话已撤销
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InvalidCredentials:
      description: 用户名或密码错误；不存在账号与错误密码使用相同外部响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InvalidResetGrant:
      description: 重置授权无效、过期或已使用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    AccountLocked:
      description: 账号已锁定；不返回账号存在性的额外信息
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    AccessDenied:
      description: 已认证但动作、数据范围、记录状态或会话条件禁止
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ResourceNotAvailable:
      description: 对象不存在或主体无权读取，响应不区分原因
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ValidationFailed:
      description: 请求字段、受控枚举或业务校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 状态、范围、生效期、冻结保护或唯一性冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    TooManyRequests:
      description: 触发登录或恢复请求频率限制
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...